test: remove unneeded triggerKeyboardEvent and inline getEditableElement
